Перенес все, что касается блока users и subs в приложение users
