--- backend/api/serializers.py
@@ -1,53 +1,0 @@
-import base64
-
-from django.core.files.base import ContentFile
-from djoser.serializers import UserSerializer as DjoserUserSerializer
-from rest_framework import serializers
-from users.models import Subscription, User
-
-
-class Base64ImageField(serializers.ImageField):
-    def to_internal_value(self, data):
-        if isinstance(data, str) and data.startswith('data:image'):
-            format, imgstr = data.split(';base64,')
-            ext = format.split('/')[-1]
-
-            data = ContentFile(base64.b64decode(imgstr), name='temp.' + ext)
-
-        return super().to_internal_value(data)
-
-
-class AvatarSerializer(serializers.Serializer):
-    avatar = Base64ImageField(required=False)
-
-
-class CurrentUserSerializer(DjoserUserSerializer):
-    is_subscribed = serializers.SerializerMethodField()
-
-    class Meta:
-        model = User
-        fields = (
-            'id',
-            'email',
-            'username',
-            'first_name',
-            'last_name',
-            'avatar',
-            'is_subscribed',
-        )
-
-    def get_is_subscribed(self, obj):
-        return False
-
-
-class UserSerializer(CurrentUserSerializer):
-    def get_is_subscribed(self, obj):
-        request = self.context['request']
-        try:
-            Subscription.objects.get(
-                follower_id=request.user.id,
-                followed_id=obj.id
-            )
-            return True
-        except Subscription.DoesNotExist:
-            return False
