+++ backend/api/urls.py
@@ -1,7 +1,7 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from .views import UserViewSet
+from users.views import UserViewSet
 
 api_v1 = DefaultRouter()
 api_v1.register(r'users', UserViewSet, basename='users')
